extract some common components based on the design system
